Move settings list to a new file
