+++ Gui.py
@@ -15,7 +15,8 @@ from GuiUtils import ToolTips, set_icon, BackgroundTask, BackgroundTaskProgress,
 from Main import main
 from Utils import is_bundled, local_path, default_output_path, open_file, check_version
 from Patches import get_tunic_color_options, get_navi_color_options
-from Settings import Settings, setting_infos
+from Settings import Settings
+from SettingsList import setting_infos
 from version import __version__ as ESVersion
 import webbrowser
 import WorldFile
