+++ version.py
@@ -1,1 +1,1 @@
-__version__ = '3.3.3 f.LUM'
+__version__ = '3.3.4 f.LUM'
